* 'master' of github.com:erikdoe/ocmock:
  silence Xcode 11.4 warnings w.r.t. upgrade
  Remove unneeded dealloc from OCPartialMockObject
  Clean up descriptions.
  Allow expectation to complete in test
  Small refactoring and some formatting changes.
  Catches references to self when not using macros.
  Cast mock arguments in OCMVerify* macros
  Fix up duplicate header copy
  Fix up printf specifiers
  Fix for #385.
  Added missing header to podspec.

+++ Source/OCMock/OCMQuantifier.m
@@ -79,7 +79,7 @@
     {
         case 0:  return @"never";
         case 1:  return @"once";
-        default: return [NSString stringWithFormat:@"%ld times", expectedCount];
+        default: return [NSString stringWithFormat:@"%lu times", (unsigned long)expectedCount];
     }
 }
